% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBPV.R
\name{getBPV}
\alias{getBPV}
\title{getBPV(y, face, couponRate, m,  ppy=1, dy=.0001)}
\usage{
getBPV(y, face, couponRate, m, ppy = 1, dy = 1e-04)
}
\arguments{
\item{y}{- yield for ppy compounding}

\item{face}{- face amount of bond}

\item{couponRate}{- coupan rate}

\item{m}{- number of coupon periods}

\item{ppy}{- payments per year 1= annual, 2= semi, etc}

\item{dy}{- yield perturbation, default is .0001}
}
\value{
change in bond price due to yield change of dy
}
\description{
getBPV(y, face, couponRate, m,  ppy=1, dy=.0001)
}
\examples{

y = 0.03
face = 2000000
couponRate = 0.04
m = 10
ppy = 2

getBPV(y, face, couponRate, m,  1)

}
